Update label encoding comment
Label encoding is way weirder than suspected. For starters, it appears
that domain labels are in practice restricted to ASCII; unicode ones use
punycode, not UTF-8. However, nothing in the DNS spec suggests any
encoding, and it seems to apply that _any_ bytes would be acceptable,
including printable ASCII characters that shouldn't be allowed in a
label (like "."), ASCII control characters, and characters outside of
ASCII.