Use SO_REUSEPORT to give each thread an independent socket
This code now is platform-specific, which could be fixed by handling
systems that don't support SO_REUSEPORT, but gets rid of the shared
memory references to a long-lived UDP socket, which I rather like. It
should be easy to support this behavior on Windows as well (it's the
default), but non-Unix and non-Windows platforms would possibly need to
fall back on the shared socket approach.

For a comprehensive overview of differences between platforms, see
https://stackoverflow.com/questions/14388706/how-do-so-reuseaddr-and-so-reuseport-differ

The net2 crate appears abandoned but did support a more abstract
interface into this.